+++ src/jass/game_objects/play/healthbar.h
@@ -23,7 +23,7 @@ class Healthbar : public GameObject {
   void Create(void) override;
   void Start(void) override;
 
-  void Update(const uint32_t dt) override;
+  void Update(const double dt) override;
   void Render(void) override;
 
  private:
